+++ examples/hello_world.py
@@ -10,7 +10,7 @@ async def fetch_me() -> None:
 
     # returns all Destiny 2 available memberships for the player.
     async with client.rest:
-        memberships = await client.fetch_player("Fate怒", 4275)
+        memberships = await client.fetch_membership("Fate怒", 4275)
 
         # Iterate over all memberships.
         for membership in memberships: