+++ tests/aiobungie/test_client.py
@@ -83,10 +83,8 @@ async def test_fetch_clan_members():
     assert len(ms) == 1
     for member in ms:
         assert isinstance(member, aiobungie.crate.ClanMember)
-        assert isinstance(member.bungie, aiobungie.crate.PartialBungieUser)
-        assert member.bungie.name == "Fate怒"
-        fetched_user = await member.bungie.fetch_self()
-        assert isinstance(fetched_user, aiobungie.crate.BungieUser)
+        assert isinstance(member.bungie_user, aiobungie.crate.PartialBungieUser)
+        assert member.bungie_user.name == "Fate怒"
 
 
 async def test_fetch_inventory_item():
@@ -97,12 +95,10 @@ async def test_fetch_inventory_item():
 async def test_fetch_app():
     a = await client.fetch_application(33226)
     assert isinstance(a, aiobungie.crate.Application)
-    fetched_user = await a.owner.fetch_self()
-    assert isinstance(fetched_user, aiobungie.crate.BungieUser)
 
 
 async def test_player():
-    p = await client.fetch_player("Fate怒", 4275)
+    p = await client.fetch_membership("Fate怒", 4275)
     profile = await p[0].fetch_self_profile([aiobungie.ComponentType.PROFILE])
     assert isinstance(profile, aiobungie.crate.Component)
     assert isinstance(profile.profiles, aiobungie.crate.Profile)
@@ -125,11 +121,6 @@ async def test_fetch_character():
     assert c.progressions
     assert c.profile_records is None
     assert c.item_components
-    acts = await c.character.fetch_activities(aiobungie.GameMode.RAID, limit=10)
-    assert len(acts) == 10
-    for act in acts:
-        assert act.mode is aiobungie.GameMode.RAID
-        assert isinstance(act, aiobungie.crate.Activity)
 
 
 async def test_profile():
@@ -208,9 +199,9 @@ async def test_profile():
 async def test_membership_types_from_id():
     u = await client.fetch_membership_from_id(MID)
     assert isinstance(u, aiobungie.crate.User)
-    assert isinstance(u.bungie, aiobungie.crate.BungieUser)
-    for du in u.destiny:
-        assert isinstance(du, aiobungie.crate.DestinyMembership)
+    assert isinstance(u.bungie_user, aiobungie.crate.BungieUser)
+    for mem in u.memberships:
+        assert isinstance(mem, aiobungie.crate.DestinyMembership)
 
 
 async def test_search_users():